Added ScrollContainer to TokenList
