+++ src/components/TokenList/TokenList.styles.tsx
@@ -1,35 +1,11 @@
-import styled, { css } from "styled-components/macro";
+import styled from "styled-components/macro";
 
 import isActiveLanguageLogographic from "../../helpers/isActiveLanguageLogographic";
 import { InputTextStyle } from "../../style/mixins";
 import { fontWide } from "../../style/themes";
 import Icon from "../Icon/Icon";
-import { ScrollContainer } from "../ModalOverlay/ModalOverlay.styles";
 import TextInput from "../TextInput/TextInput";
 import { StyledInput } from "../TextInput/TextInput.styles";
-
-type StyledScrollContainerProps = {
-  $overflow?: boolean;
-  $scrolledToBottom?: boolean;
-};
-
-export const StyledScrollContainer = styled(
-  ScrollContainer
-)<StyledScrollContainerProps>`
-  position: relative;
-  margin-block-start: 0.625rem;
-  margin-inline-start: -0.875rem;
-  width: calc(100% + 3.25rem);
-  max-height: 20rem;
-  padding-inline: 0.875rem 2.25rem;
-  padding-block-start: 0.125rem;
-
-  ${(props) =>
-    props.$scrolledToBottom &&
-    css`
-      -webkit-mask-image: none;
-    `}
-`;
 
 export const ContentContainer = styled.div`
   position: relative;
